[UPD] moved login screen in own seperate file to not overclutter main.dart
