+++ client/apk/lib/main.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:google_fonts/google_fonts.dart';
-import 'screens/register_screen.dart';
+import 'screens/login_screen.dart';
 import 'screens/home_screen.dart';
-import 'services/api_service.dart';
 import 'services/storage_service.dart';
 
 void main() {
@@ -69,245 +67,6 @@ class _SplashScreenState extends State<SplashScreen> {
           valueColor: AlwaysStoppedAnimation<Color>(Colors.white),
         ),
       ),
-    );
-  }
-}
-
-class LoginScreen extends StatefulWidget {
-  @override
-  _LoginScreenState createState() => _LoginScreenState();
-}
-
-class _LoginScreenState extends State<LoginScreen> {
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _passwordController = TextEditingController();
-  bool _isLoading = false;
-  String? _errorMessage;
-
-  Future<void> _handleLogin() async {
-    setState(() {
-      _isLoading = true;
-      _errorMessage = null;
-    });
-
-    final email = _emailController.text.trim();
-    final password = _passwordController.text;
-
-    if (email.isEmpty || password.isEmpty) {
-      setState(() {
-        _isLoading = false;
-        _errorMessage = 'Please enter email and password';
-      });
-      return;
-    }
-
-    final result = await ApiService.login(email, password);
-
-    setState(() {
-      _isLoading = false;
-    });
-
-    if (result['success']) {
-      // Save token and user data
-      await StorageService.saveToken(result['token']);
-      final user = result['user'];
-      await StorageService.saveUserData(user['username'], user['email']);
-
-      // Navigate to home
-      if (mounted) {
-        Navigator.pushReplacement(
-          context,
-          PageRouteBuilder(
-            pageBuilder: (context, animation, secondaryAnimation) => HomeScreen(),
-            transitionDuration: Duration.zero,
-            reverseTransitionDuration: Duration.zero,
-          ),
-        );
-      }
-    } else {
-      setState(() {
-        _errorMessage = result['error'];
-      });
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.black,
-      resizeToAvoidBottomInset: false,
-      body: SafeArea(
-        child: Column(
-          children: [
-            Expanded(
-              child: SingleChildScrollView(
-                child: Padding(
-                  padding: EdgeInsets.symmetric(horizontal: 40, vertical: 40),
-                  child: Column(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      // AREA logo
-                      Transform.scale(
-                        scale: 1.5,
-                        child: Container(
-                          padding: EdgeInsets.symmetric(horizontal: 5, vertical: 5),
-                          decoration: BoxDecoration(
-                            border: Border.all(color: Colors.white, width: 3),
-                          ),
-                          child: Text(
-                            'AR\nEA',
-                            style: GoogleFonts.russoOne(
-                              fontSize: 48,
-                              color: Colors.white,
-                              fontWeight: FontWeight.bold,
-                              height: 0.8,
-                              letterSpacing: 5,
-                            ),
-                          ),
-                        ),
-                      ),
-                      SizedBox(height: 40),
-
-                      // Email field
-                      TextField(
-                        controller: _emailController,
-                        style: TextStyle(color: Colors.white),
-                        keyboardType: TextInputType.emailAddress,
-                        decoration: InputDecoration(
-                          hintText: 'Email',
-                          hintStyle: TextStyle(color: Colors.grey),
-                          enabledBorder: UnderlineInputBorder(
-                            borderSide: BorderSide(color: Colors.white),
-                          ),
-                          focusedBorder: UnderlineInputBorder(
-                            borderSide: BorderSide(color: Colors.white),
-                          ),
-                        ),
-                      ),
-                      SizedBox(height: 20),
-
-                      // Password field
-                      TextField(
-                        controller: _passwordController,
-                        obscureText: true,
-                        style: TextStyle(color: Colors.white),
-                        decoration: InputDecoration(
-                          hintText: 'Password',
-                          hintStyle: TextStyle(color: Colors.grey),
-                          enabledBorder: UnderlineInputBorder(
-                            borderSide: BorderSide(color: Colors.white),
-                          ),
-                          focusedBorder: UnderlineInputBorder(
-                            borderSide: BorderSide(color: Colors.white),
-                          ),
-                        ),
-                      ),
-                      SizedBox(height: 10),
-
-                      // Error message
-                      if (_errorMessage != null)
-                        Padding(
-                          padding: EdgeInsets.only(bottom: 10),
-                          child: Text(
-                            _errorMessage!,
-                            style: TextStyle(color: Colors.red, fontSize: 14),
-                            textAlign: TextAlign.center,
-                          ),
-                        ),
-                      SizedBox(height: 20),
-
-                      // OAuth icons (horizontal)
-                      Row(
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          MouseRegion(
-                            cursor: SystemMouseCursors.click,
-                            child: GestureDetector(
-                              onTap: () {
-                                print('Google login pressed');
-                              },
-                              child: Image.asset(
-                                'assets/google-icon.png',
-                                width: 50,
-                                height: 50,
-                                filterQuality: FilterQuality.high,
-                              ),
-                            ),
-                          ),
-                          SizedBox(width: 20),
-                          MouseRegion(
-                            cursor: SystemMouseCursors.click,
-                            child: GestureDetector(
-                              onTap: () {
-                                print('Facebook login pressed');
-                              },
-                              child: Image.asset(
-                                'assets/facebook-icon.png',
-                                width: 50,
-                                height: 50,
-                                filterQuality: FilterQuality.high,
-                              ),
-                            ),
-                          ),
-                        ],
-                      ),
-                      SizedBox(height: 30),
-
-                      // Login button
-                      ElevatedButton(
-                        onPressed: _isLoading ? null : _handleLogin,
-                        style: ElevatedButton.styleFrom(
-                          backgroundColor: Colors.white,
-                          foregroundColor: Colors.black,
-                          minimumSize: Size(double.infinity, 50),
-                        ),
-                        child: _isLoading
-                            ? SizedBox(
-                                height: 20,
-                                width: 20,
-                                child: CircularProgressIndicator(
-                                  strokeWidth: 2,
-                                  valueColor: AlwaysStoppedAnimation<Color>(Colors.black),
-                                ),
-                              )
-                            : Text('Login', style: TextStyle(fontSize: 18)),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-      Padding(
-        padding: EdgeInsets.only(bottom: 20),
-        child: MouseRegion(
-          cursor: SystemMouseCursors.click,
-          child: GestureDetector(
-            onTap: () {
-              Navigator.push(
-                context,
-                PageRouteBuilder(
-                  pageBuilder: (context, animation, secondaryAnimation) => RegisterScreen(),
-                  transitionDuration: Duration.zero,
-                  reverseTransitionDuration: Duration.zero,
-                ),
-              );
-            },
-            child: Text(
-              'Create a Account | Forgot password',
-              style: TextStyle(
-                color: Colors.grey,
-                fontSize: 12,
-                fontWeight: FontWeight.w300,
-                decoration: TextDecoration.underline,
-                decorationColor: Colors.grey.withOpacity(0),
-              ),
-            ),
-          ),
-        ),
-      ),
-    ],
-  ),
-),
     );
   }
 }
