+++ client/apk/lib/screens/home_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import '../services/storage_service.dart';
-import '../main.dart';
+import 'login_screen.dart';
 
 class HomeScreen extends StatelessWidget {
   Future<void> _handleLogout(BuildContext context) async {
